Cleanup Code, Edit EMC values

+++ src/main/java/io/github/chaosawakens/common/blocks/AntInfestedOre.java
@@ -1,7 +1,5 @@
 package io.github.chaosawakens.common.blocks;
 
-import java.util.function.Supplier;
-
 import net.minecraft.block.BlockState;
 import net.minecraft.block.OreBlock;
 import net.minecraft.enchantment.EnchantmentHelper;
@@ -13,18 +11,20 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.GameRules;
 import net.minecraft.world.server.ServerWorld;
 
+import java.util.function.Supplier;
+
 public class AntInfestedOre extends OreBlock {
-	
-	private final Supplier<? extends EntityType<?>> ant;
-	
+
+    private final Supplier<? extends EntityType<?>> ant;
+
     public AntInfestedOre(Supplier<? extends EntityType<?>> ant, Properties properties) {
         super(properties);
         this.ant = ant;
     }
 
     private void spawnAnt(ServerWorld world, BlockPos pos) {
-    	MonsterEntity antEntity =  (MonsterEntity) ant.get().create(world);
-    	antEntity.setLocationAndAngles((double)pos.getX() + 0.5D, (double)pos.getY(), (double)pos.getZ() + 0.5D, 0.0F, 0.0F);
+        MonsterEntity antEntity = (MonsterEntity) ant.get().create(world);
+        antEntity.setLocationAndAngles((double) pos.getX() + 0.5D, pos.getY(), (double) pos.getZ() + 0.5D, 0.0F, 0.0F);
         world.addEntity(antEntity);
         antEntity.spawnExplosionParticle();
     }
